[REMOVE] Some lines of code were causing bugs (v1.0.5)
In one of the if statements, one of the conditions caused a bug that made the game literally unplayable. Also removed a line of code in the CSS I probably made while drunk because I have no clue what it was doing there. AND changed the commentary from Dutch to English (because I'm super international now ;D)